update to nav style
