+++ src/components/svg/SiteName.jsx
@@ -1,6 +1,7 @@
 function TitleSvg(props) {
   return (
     <svg
+      className={`${props.styles}`}
       xmlns="http://www.w3.org/2000/svg"
       xmlnsXlink="http://www.w3.org/1999/xlink"
       width={241}
